# src/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: "MyMediathek"
  description: "Interface to synchronize and provide bookmarked movies and handle movies."
basePath: "/api"
...
schemes:
- "http"
- "https"
paths:
...
  /bookmarks:
    post:
      tags:
      - "Bookmarks"
      summary: "Add new bookmark(s)"
      operationId: "server.bookmark.bookmarks_post"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "bookmarks"
        description: "Array of bookmark object(s) to be added to the database.<br>If an element already exists it will be overwritten if the modification date is older, otherwise the entry is ignored.<br>If mandatory elements are missing, an error is returned."
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/bookmark"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              nb:
                description: "Number of added bookmarks"
                type: "integer"
              detail:
                type: "array"
                items:
                  properties:
                    id:
                      type: "integer"
                      format: "int64"
                      description: "Bookmark id from request. Java hash value of sender and url:<br>(java) int id = (sender + url).hashCode().abs()"
                    expiry:
                      type: "integer"
                      description: "Expiry time stamp in seconds since since Jan 01 1970. If no expiry is set the value is 2000000000."
        "400":
          description: "Invalid request, e.g. body syntax wrong"
        "422":
          description: "Invalid data , e.g. invalid date like 0.20.2020"

    get:
      tags:
      - "Bookmarks"
      summary: "Retrieve bookmarks"
      description: "If no bookmarks are found an empty response is returned.<br><br>Optional filters:<ul><li>A category name can be specified as query parameter (e.g. '/bookmarks?category=China'), to retrieve only bookmarks with the requested category. If the category does not exit an error is returned, if no bookmarks with the category exist, an empty list is returned.</li><li>Sender name: It can be filtered with a optional sender name.</li></ul><b>Note: the filters can not be combined!</b><br><br>The optional sort parameter can be specified to force a sort, the direction can be inverted by setting the sub parameter desc."
      operationId: "server.bookmark.bookmarks_get"
      parameters:
      - name: "category"
        in: "query"
        description: "optional bookmark category, to retrieve movies without category use special value '@NULL'"
        type: "string"
      - name: "sender"
        in: "query"
        description: "optional sender name, to retrieve only bookmarks for a specific sender"
        type: "string"
      - name: "sort"
        in: "query"
        description: "Sort order (optional)<br>
          Possible values are<br>
          * date: Sort by send date (and title)<br>
          * duration: Sort by duration (and title)<br>
          * expiry: Sort by expiry date (and title)<br>
          * theme: Sort by theme (and title)<br>
          * title: Sort by title"
        type: "string"
        enum:
          - date
          - duration
          - expiry
          - theme
          - title
      - name: "desc"
        in: "query"
        description: "sort direction (only valid with sort)"
        type: "boolean"
      - name: "limit"
        in: "query"
        description: "number of records to return"
        type: "integer"
      - name: "offset"
        in: "query"
        description: "offset for first value to return "
        type: "integer"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/bookmark"
        "404":
          description: "Invalid query parameter"
  /bookmarks/delete:
    post:
      tags:
      - "Bookmarks"
      summary: "Delete bookmark(s) with given ids"
      operationId: "server.bookmark.bookmarks_delete"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "ids"
        description: "Array of bookmark id(s) to be removed from the database"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/id"
      responses:
        "204":
          description: "Successful operation"
        "400":
          description: "Not supported body"
        "404":
          description: "No records deleted"
        "500":
          description: "Intenal error"
  /bookmarks/{hid}:
    delete:
      tags:
      - "Bookmarks"
      summary: "Delete given bookmark"
      description: ""
      operationId: "server.bookmark.bookmarksByIdDelete"
      parameters:
      - name: "hid"
        in: "path"
        description: "bookmark id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
                description: "Bookmark id from request. Java hash value of sender and url:<br>(java) int id = (sender + url).hashCode().abs()"
              deleted:
                type: boolean
                description: True, means record was deleted
            example:
              { 'id': 1317922529 , 'deleted': true}

        "404":
          description: "bookmark not found"

    get:
      tags:
      - "Bookmarks"
      summary: "Get bookmark"
      description: "Get data for given bookmark."
      operationId: "server.bookmark.bookmarksByIdGet"
      produces:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "ID of the bookmark"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/bookmark"
        "404":
          description: "bookmark not found"
        "422":
          description: "value error"

    patch:
      tags:
      - "Bookmarks"
      summary: "Set additional bookmark information"
      description: "Set information parameters (e.g. category, expiry date) for given bookmark id."
      operationId: "server.bookmark.bookmarksByIdPatch"
      produces:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "ID of the bookmark to update"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "patches"
        description: "Array of JSON patch definitions (RFC6902)"
        schema:
          type: "array"
          items:
            type: object
            properties:
              op:
                type: string
                enum: [replace, remove]
                description: "patch operation, only replace and remove are supported"
              name:
                type: string
                enum: [category, imgurl, videoformat, expiry]
                description: "parameter name to be modified"
              value:
                type: string
                description: "new parameter value, empty for remove"

      responses:
        "200":
          description: "successful operation"
          schema:
            type: object
            properties:
              id:
                type: integer
              success:
                type: boolean
              errmsg:
                type: string
        "404":
          description: "bookmark not found"
        "422":
          description: "Invalid data , e.g. invalid date like 0.20.2020"
        "500":
          description: "player error, e.g. no player configured"
  /bookmarks/{hid}/seen:
    patch:
      tags:
      - "Bookmarks"
      summary: "Set bookmark seen information"
      description: "Set bookmark seen state to 'True' or 'False'"
      operationId: "server.bookmark.bookmarksByIdSeenPatch"
      produces:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "ID of the bookmark to update"
        required: true
        type: "integer"
        format: "int64"
      - name: "seen"
        in: "query"
        description: "seen state"
        required: true
        type: "boolean"

      responses:
        "204":
          description: "successful operation"
  /bookmarks/{hid}/valid:
    patch:
      tags:
      - "Bookmarks"
      summary: "Set bookmark to valid or invalid"
      description: "Set bookmark valid state to 'True' or 'False'"
      operationId: "server.bookmark.bookmarksByIdValidPatch"
      produces:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "ID of the bookmark to update"
        required: true
        type: "integer"
        format: "int64"
      - name: "valid"
        in: "query"
        description: "valid state"
        required: true
        type: "boolean"

      responses:
        "204":
          description: "successful operation"
  /bookmarks/{hid}/expiry:
    get:
      tags:
      - "Bookmarks"
      summary: "Get bookmark expiry date from the webpage."
      description: "Get expiry date for given bookmark id from the webpage."
      operationId: "server.bookmark.bookmarksByIdExpiryGet"
      produces:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "ID of the bookmark"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              expiry:
                type: "integer"
                description: "Expiry time stamp in seconds since since Jan 01 1970."
            example:
             {"expiry" : 1636758000 }

        "404":
          description: "bookmark not found"
        "422":
          description: "value error"
...
definitions:
  bookmark:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Java hash value of sender and url:<br>(java) int id = (sender + url).hashCode().abs()"
      modified:
        type: "integer"
        description: "Last modification time stamp in format 'yyyymmddhhmmss'"
      sender:
        type: "string"
      thema:
        type: "string"
        x-nullable: true
      titel:
        type: "string"
        x-nullable: true
      category:
        type: "string"
        x-nullable: true
      url:
        type: "string"
      description:
        type: "string"
        x-nullable: true
      duration:
        type: "integer"
        description: "movie length in seconds"
      sendDate:
        type: "string"
        x-nullable: true
      seen:
        type: "boolean"
        default: false
      expiry:
        type: "integer"
      note:
        type: "string"
        x-nullable: true
      website:
        type: "string"
        x-nullable: true
        description: "Url of mediathek website for this movie"
      imgurl:
        type: "string"
        x-nullable: true
        description: "movie image URL in mediathek"
      videoformat:
        type: "string"
        x-nullable: true
        description: "video format e.g. 720p, 1080p"
...
  id:
    type: "integer"
    format: "int64"
    description: "Unique movie id. Create as Java hash value of sender and url:<br>(java) int id = (sender + url).hashCode().abs()"
    example: 26071963